Add db migration running on startup
