+++ sql/V0_3_0__Add_hashes_table.sql
